form/requirement.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
blinker==1.7.0
Brotli==1.1.0
bs4==0.0.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cssselect2==0.7.0
Django==4.2.5
django-phonenumber-field==7.1.0
Flask==3.0.0
fonttools==4.43.1
html5lib==1.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
phonenumbers==8.13.21
Pillow==10.1.0
psycopg2-binary==2.9.9
pycparser==2.21
pydyf==0.8.0
pyphen==0.14.0
reportlab==4.0.6
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tinycss2==1.2.1
tzdata==2023.3
urllib3==2.0.7
weasyprint==60.1
webencodings==0.5.1
Werkzeug==3.0.1
whitenoise==6.6.0
zopfli==0.2.3
